fix cart wrong product count

+++ carts/models.py
@@ -2,6 +2,8 @@ from _decimal import Decimal
 from django.contrib.auth.models import User
 from django.core.exceptions import ValidationError
 from django.db import models
+from django.db.models.signals import post_save, pre_save
+from django.dispatch import receiver
 
 from orders.models import Order
 from products.models import Product
@@ -20,12 +22,19 @@ class Cart(models.Model):
     def __str__(self):
         return str(self.id)
 
-    def save(self, *args, **kwargs):
-        if self.product is not None:
+    def save(self, delete_old_cart=True, *args, **kwargs):
+        if delete_old_cart is True and self.product is not None:
             max_quantity = self.product.available_quantity
             try:
-                prev_unordered_cart = Cart.objects.get(user=self.user, product=self.product, ordered=False)
-                if prev_unordered_cart.quantity + self.quantity <= max_quantity:
+                prev_unordered_cart = Cart.objects.filter(user=self.user,
+                                                          product=self.product,
+                                                          ordered=False).first()
+                if prev_unordered_cart is None:
+                    self.clean_total()
+                    return super(Cart, self).save(*args, **kwargs)
+                elif prev_unordered_cart.id == self.id:
+                    return
+                elif prev_unordered_cart.quantity + self.quantity <= max_quantity:
                     self.quantity += prev_unordered_cart.quantity
                     # self.total += self.product.price_taka * self.quantity
                     prev_unordered_cart.delete()
@@ -33,16 +42,21 @@ class Cart(models.Model):
                     raise ValidationError
             except Cart.DoesNotExist:
                 pass
-            try:
-                self.total += Decimal(self.product.price_taka * self.quantity)
-            except TypeError as t:
-                self.total += float(self.product.price_taka * self.quantity)
 
+        self.clean_total()
         return super(Cart, self).save(*args, **kwargs)
+
+    def clean_total(self):
+        try:
+            self.total += Decimal(self.product.price_taka * self.quantity)
+        except TypeError as t:
+            self.total += float(self.product.price_taka * self.quantity)
 
     @property
     def seller(self):
         return self.product.seller
+
+
